added function that abstracts the call logic to fetch images and cache them
