+++ RickAndMortyApp.xcodeproj/xcuserdata/User.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "D0C35CFB-E9B8-4251-8C26-FAE55E320D0C"
+   type = "1"
+   version = "2.0">
+</Bucket>
