+++ RickAndMortyApp/ViewModels/CharacterCollectionViewCelltViewViewModel.swift
@@ -34,17 +34,8 @@ final class CharacterCollectionViewCelltViewViewModel: Hashable, Equatable {
             completion(.failure(URLError(.badURL)))
             return
         }
-        let request = URLRequest(url: url)
         
-        let task = URLSession.shared.dataTask(with: request) { data, _, error in
-            guard let data = data, error == nil else {
-                completion(.failure(error ?? URLError(.badServerResponse)))
-                return
-            }
-            
-            completion(.success(data))
-        }
-        task.resume()
+        AppImageLoader.shared.downloadImage(url, completion: completion)
     }
     
     // MARK: - Hashable
